### 2016-05-17 #22 @whedon
[REVIEW]: GRUPO: Gauging Research University Publication Output
**Submitting author:** @vpnagraj (<a href="http://orcid.org/0000-0003-0060-566X">VP Nagraj</a>)
**Repository:** <a href="https://github.com/vpnagraj/grupo" target ="_blank">https://github.com/vpnagraj/grupo</a>
**Version:** v1.0.0
**Archive:** <a href="http://dx.doi.org/10.6084/m9.figshare.3383653.v1" target ="_blank">https://dx.doi.org/10.6084/m9.figshare.3383653.v1</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.6084/m9.figshare.3383653.v1">software archive</a> resolve?
- [ ] **Performance:** Have the performance claims of the software been confirmed?
- [ ] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

Compiled paper PDF: [10.21105.joss.00022.pdf](https://github.com/openjournals/joss-reviews/files/269280/10.21105.joss.00022.pdf)
