### 2016-05-17 #22 @janschulz
I've had a short look at the software: it is a one page shiny app where you can chose two US universities and a timerange and the app displays the number of papers in PUBMED.

Regarding `Have the functional claims of the software been confirmed?`: IMO yes, but there is IMO a lot more things to build around the idea of analysing such data than simply choosing a date range and showing a bar plot (e.g. numbers over years).

There are no tests.
