### 2016-05-17 #22 @aespinosa
# General Checks

* version - there is no github release or git tag for v1.0.0

# Documetation

* example usage - there's no documented example usage. a screenshot on how to use would be nice
* automated tests - there are none. maybe take a look at Selenium to do basic verification of certain gui elments
* community guidelines - there is no CONTRIBUTING.md file or equivalent

# Software paper

* statement of need - The summary should focus more on the motivation with building the software rather that how it works.

# Reviewer conclusion

I agree with @janschulz that more work to do more work of analyzing the data and have  greater justification on its contribution before being accepted.
