### 2016-05-18 #22 @vpnagraj
@arfon @aespinosa @janschulz

Thank you all for taking the time to review the submission.

I will take a step back and try to address the issues around functionality / research applications for this software.

Is there currently a process in place for re-submitting? How long will the paper remain under review?

Thanks again for the feedback.
