### 2016-05-23 #22 @arfon
> I will take a step back and try to address the issues around functionality / research applications for this software.

Excellent.

> Is there currently a process in place for re-submitting? How long will the paper remain under review?

We're happy to leave this pending until you've made your improvements to the package. Please comment on this thread when you're ready for a re-review.
