### 2018-01-08 #22 @arfon
@vpnagraj - have you had a chance to take a look at making improvements to this submission? Also, if you're not interested in pursuing this submission to JOSS any longer we can withdraw your submission,.
