### 2018-01-08 #22 @arfon
👍 ok thanks @vpnagraj - I'll go ahead and withdraw the submission on your behalf. Thanks for you interest in JOSS :-)
